+++ library/normal_girl_class.hpp
@@ -14,8 +14,8 @@ namespace data {
             std::string commit_to;
 
         public:
-           normal_girl(std::string name_id, int attractiveness, int maintainance_cost,
-                 int iq, char criteria);
+            normal_girl(std::string name_id, int attractiveness, 
+                        double maintainance_cost, int iq, char criteria);
             void set_happiness(int happiness);
             void change_status(char status);
             void change_status(char status, std::string commit_to);
